feat: tcp mainstream
- model 수정해서 tcp로 받은 정보 그대로 모델에 들어가도록
