+++ .gitignore
@@ -1,3 +1,8 @@
 model/pretrained_models
+model/output/
+*.pyc
+*.*.pyc
+*.png
 py_model/dat/shape_predictor_81_face_landmarks.dat
 py_model/util/test.jpg
+
