+++ model/utils/tcp_server.py
@@ -3,21 +3,26 @@ import cv2
 import threading
 import numpy
 import pybase64 as base64
+from PIL import Image
 
 class ServerSocket:
-    def __init__(self, ip, port):
+    def __init__(self, ip='172.17.0.4', port=8080):
         self.TCP_IP = ip
         self.TCP_PORT = port
         self.socketOpen()
-        self.receiveThread = threading.Thread(target=self.receiveImages)
-        self.receiveThread.start()
-
+        self.img = self.receiveImages()
+        self.socketClose()
+        # self.receiveThread = threading.Thread(target=self.receiveImages)
+        # self.receiveThread.start()
+        # not multi thread
+                
     def socketClose(self):
         self.sock.close()
         print(u'Server socket [ TCP_IP: ' + self.TCP_IP + ', TCP_PORT: ' + str(self.TCP_PORT) + ' ] is close')
 
     def socketOpen(self):
         self.sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+        self.sock.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
         self.sock.bind((self.TCP_IP, self.TCP_PORT))
         self.sock.listen(1)
         print(u'Server socket [ TCP_IP: ' + self.TCP_IP + ', TCP_PORT: ' + str(self.TCP_PORT) + ' ] is open')
@@ -26,20 +31,22 @@ class ServerSocket:
 
     def receiveImages(self):
         try:
-            while True:
+            # while True:
+            for i in range(1):
                 length = self.recvall(self.conn, 64)
                 # print(length)
                 length1 = length.decode('utf-8')
-                print(length1)
+                # print(length1)
                 stringData = self.recvall(self.conn, int(length1))
                 # print("st : " , stringData)
                 data = numpy.frombuffer(base64.b64decode(stringData), dtype = numpy.uint8)
-                numpy.save("data2",data)
-                print("data",data)
-                print(data.shape)
+                # numpy.save("data2",data)
+                # print("data",data)
+                # print(data.shape)
                 decimg = cv2.imdecode(data, 1)
-                print("dec : ", decimg)
-                return decimg
+                imageRGB = cv2.cvtColor(decimg , cv2.COLOR_BGR2RGB)
+                pil_img = Image.fromarray(imageRGB, "RGB")
+            return pil_img
                 # cv2.imwrite("image.png", decimg)
                 # cv2.waitKey(1)
                 
@@ -60,7 +67,8 @@ class ServerSocket:
         return buf
 
 def main():
-    server = ServerSocket('192.168.0.100', 5050)
+    server = ServerSocket('172.17.0.4', 8080)
+    server.receiveImages()
 
 if __name__ == "__main__":
     main()
